flask_app/requirements.txt
Flask
numpy
python-dotenv
Flask-SQLAlchemy
mysql-connector-python
Flask-JWT-Extended
flask-cors
python-dotenv